+++ community/io/src/main/java/org/neo4j/io/pagecache/impl/muninn/MuninnPage.java
@@ -341,7 +341,7 @@ final class MuninnPage extends StampedLock implements Page
         FlushEvent event = flushOpportunity.beginFlush( filePageId, getCachePageId(), swapper );
         try
         {
-            int bytesWritten = swapper.write( filePageId, this );
+            long bytesWritten = swapper.write( filePageId, this );
             markAsClean();
             event.addBytesWritten( bytesWritten );
             event.done();
@@ -380,7 +380,7 @@ final class MuninnPage extends StampedLock implements Page
         // the file page, so any subsequent thread that finds the page in their
         // translation table will re-do the page fault.
         this.filePageId = filePageId; // Page now considered isLoaded()
-        int bytesRead = swapper.read( filePageId, this );
+        long bytesRead = swapper.read( filePageId, this );
         faultEvent.addBytesRead( bytesRead );
         faultEvent.setCachePageId( getCachePageId() );
         this.swapper = swapper; // Page now considered isBoundTo( swapper, filePageId )
